Deploy Lambda artifact from S3 bucket
Pull the deployment artifact from S3 using the new
terraform-external-module-artifact module. This approach has the benefit
of not checking in zip files into git, but the downside of more preamble
when wanting to test this module as per the example docs. You need to
know that the function should already be in S3 at a pre determined path.

We need to think of proper deployment and promotion of these artifacts
through different S3 buckets e.g.
artifacts.testing.cloudposse.org -> artifacts.prod.cloudposse.org